backend/requirements.txt
fastapi[standard]
requests
python-dotenv
langchain
langchain-google-genai